moved get_image_array_with_max_resolution
