+++ sciencebeam_gym/utils/cv.py
@@ -28,6 +28,36 @@ def resize_image(src: np.ndarray, width: int, height: int) -> np.ndarray:
     )
 
 
+def get_image_array_with_max_resolution(
+    image_array: np.ndarray,
+    max_width: int,
+    max_height: int
+) -> np.ndarray:
+    original_height, original_width = image_array.shape[:2]
+    if (
+        (not max_width or original_width <= max_width)
+        and (not max_height or original_height <= max_height)
+    ):
+        LOGGER.debug(
+            'image within expected dimension: %sx%s <= %sx%s',
+            original_width, original_height, max_width, max_height
+        )
+        return image_array
+    target_width_based_on_height = int(
+        original_width * max_height / original_height
+    )
+    target_height_based_on_width = int(
+        original_height * max_width / original_width
+    )
+    if max_height and (not max_width or target_width_based_on_height <= max_width):
+        return resize_image(
+            image_array, width=target_width_based_on_height, height=max_height
+        )
+    return resize_image(
+        image_array, width=max_width, height=target_height_based_on_width
+    )
+
+
 def crop_image_to_bounding_box(
     src: np.ndarray,
     bounding_box: BoundingBox,
